Enterprise of the Fallen - First version
################################
# Enterprise of the Fallen
# Original idea by: Mourn
# Code by:          Tuttu
# Writing by:       Elimdur
################################

This is the first version of the "Enterprise of the Fallen" (Temporary
name) event. At the moment, it only contains the anomaly and an event
for a successful or a failed survey. I will expand it but you can still
test it if you want. It should spawn on Tomb World only but you can use
the "add_anomaly ef_temp_category" command to spawn one. :)