Expanding the event chains
After you found the library, a new event will appear almost 5 years
later and you can have different outcomes if you are Xenophile or
Xenophobe on top of the default outcome.
This completes this anomaly as long as I don't find a good way to extend
(And not replace like suggested Kyojin) the duration of some outcomes.
If I find nothing, I will use Kyojin's workaround.