Refactor brainworm files with refactor_events.py script
This does not complete the refactor, but I did inspect the script's
changes to confirm that they were correct. There are at least a couple
of places the script did not refactor -- brainworm_triggers.txt and the
brainworm special project names. More may also exist. These are left
for a later commit.