[Star Survey] Replace incorrect Spanish with English
Some text in localisation/mem_star_survey_l_spanish.yml is no longer
correct for the current version of the event. Until translation happens
mem_star_survey.2.desc and mem_star_survey.2.a are given in English in
the Spanish localization, and mem_star_survey.2.b is using a slightly
outdated version of Spanish that is mediocre.