Fix every_owned_pop not behind hidden_effect bugs
The game can't handle running `every_owned_pop` on an entire (large)
empire if it is not inside a `hidden_effect` block. This affects
Eager Traders, Enterprise Fallen, and Music Tour. Music Tour and
Spiritualist Pilgrimage also involve `every_owned_pop` calls but only
to a single planet, so I left those unchanged; but they still result
in annoyingly large blocks, so in the long run we may want to modify
those too.