Add precursor extension/fix pilot event.
A pilot event for our project to add events to enable the precursor
event chains to be completed even if anomalies become unavailable.
See #152 for discussion of our intent.

This one allows the vultaum_1 event to spawn on a planet.

Note in the MTTH calculations that this is the MTTH to happen *per
planet*, and so (beyond 25 planets) the rate an *empire* gets the
event scales linearly with the number of planets. So an empire with
60 planets will get the events twice as often as an empire with 30
planets.