Add debug events to science convention for tracking flow problems
Added events mem_science_convention.4700, 4703, 4704, 4705, 4706,
and 4747 for debugging purposes. There is clearly some sort of flow
problem or other internal bug to the event; the code for events
mem_science_convention.3, 4, 5, and 6 clearly expects that any
attending country will receive whichever of these events corresponds
to the convention's success/failure outcome, which is only determined
once (in mem_science_convention.2). However, event 4703/4/5/6 should
be triggered many times if this happens, and it is not. Also, when the
science convention is triggered several times in a row, the countries
that do get this event will be exactly all the hosts from any
convention, which seems to indicate that 999 is not properly cleaning
convention information.

Haven't figured out why this is happening, but at least we're getting
some new data.