added country flag to mark the initial finder
should serve for follow up events after colonizing the planet. to determine if the player colonizing the planet should get some explanatory events to know what this building is about to finally be on the same information level as the initial finder.

After this start the follow up colony events that will be the same for everyone.