some small change proposals
I strongly dislike universal and absolute phrases like "All molten worlds" or "something is indestructable" as it implies that the one stating this is omniscient. A scientist would only say what he knows - which is limited by what he can observe.

I also added and changed some minor parts of the text to make it sound a little bit more believable/scientific for me. I think this just comes down to personal preference. If you don´t like it, revert it to the previous lines.