more changes to the chain desription
I want to avoid any mention of "empire" as their nation was the furthest away from being that. An empire is something based on domination and exploitation and that is not their way. Also they were never bitter or demoralized. They new their end would be coming eventually as their species had been in decline for a long time. They saw it rationally pretty much like the asgard faced their end.