primitive changes + vazuran kill switch event
noticed some strange happenings with primitive worlds having multiple event flags but those events did not happen for the player. not sure, this fixes anything, but it should not break anything either.
moonlanding now only works for planets with moons, not moons. primitives landing on a gas giant or bigger body with much higher gravity in any case would be a bit strange