Update mem_spiritualists_pilgrimage.txt
there was an issue for a steam user where some empire got targeted that should not have been possible. maybe the event target did not make it through so we try it with a global event target for now. should not do any damage as only one instance of this event can ever take place at the same time.